Added File Handler method calls in RStarTree, Node and Record classes.

+++ src/tree/Record.java
@@ -3,17 +3,17 @@ package tree;
 import java.io.Serializable;
 
 public class Record implements Serializable {
-    private int id;
+    private long id;
     private String name;
     private double[] coordinates;
 
-    public Record(int id, String name, double[] coordinates) {
+    public Record(long id, String name, double[] coordinates) {
         this.id = id;
         this.name = name;
         this.coordinates = coordinates;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
